requirements.txt
requests==2.26.0
pytz==2021.3
python-dotenv==1.0.0